pyproject.toml
[project]
name = "user-engagement-api"
version = "0.1.0"
description = ""
authors = [
    { name = "clasterpack", email = "clasterpack@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "python-logstash",
    "python3-logstash (>=0.4.80,<0.5.0)",
    "httpx (>=0.28.1,<0.29.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.venv
  | \__pycache__
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
wrap_sections = []